move play to actions
