+++ tests/arrange.py
@@ -2,12 +2,11 @@
 
 from typing import Callable, Optional
 
+from hundredandten.actions import Discard, Play
 from hundredandten.constants import (AnyStatus, BidAmount, GameRole,
                                      GameStatus, RoundStatus, SelectableSuit)
-from hundredandten.actions import Discard
 from hundredandten.game import Game
 from hundredandten.group import Group, Person
-from hundredandten.trick import Play
 
 
 def game(
