+++ tests/game/test_play_card.py
@@ -1,11 +1,11 @@
 '''Test behavior of the Game when playing a card'''
 from unittest import TestCase
 
+from hundredandten.actions import Play
 from hundredandten.constants import (HAND_SIZE, CardNumber, RoundStatus,
                                      SelectableSuit)
 from hundredandten.deck import Card
 from hundredandten.hundred_and_ten_error import HundredAndTenError
-from hundredandten.trick import Play
 from tests import arrange
