+++ tests/trick/test_trick.py
@@ -1,9 +1,10 @@
 '''Test behavior of a trick'''
 from unittest import TestCase
 
+from hundredandten.actions import Play
 from hundredandten.constants import CardNumber, SelectableSuit
 from hundredandten.deck import Card
-from hundredandten.trick import Play, Trick
+from hundredandten.trick import Trick
 
 
 class TestTrick(TestCase):
